Add buttons on top of image and make image act as background
Updated the HTML and CSS code to add two buttons on top of the image and make the image act as a background for the buttons. Used a new div with the class 'img-container' to wrap around the image and positioned the buttons within a new div with the class 'button-container'. Added CSS properties to style the buttons and ensure that the image scales proportionally.